refactor(convert): move list_item.rs tests into home module
